# perf: add microbenchmark between column-aware encoding and value encoding